Upgrade para a versão 1.5
Os arquivos foram unificados no organizer.php.|

Interfaces novas foram adicionados.

Métodos declarados no Objeto Control.

A estrurtura de If da aplicação para gerir e organizar os arquivos foram diminuidos para melhor entender.
